build(database and table): bREAKING-CHANGE: created user and url databases and view tables
Created user register and login ui and database/tables and a new table that contains the URL added.
In ui was added a table that shows every URL added

BREAKING CHANGE: Created user register and login ui and database/tables and a new table that
contains the URL added. In ui was added a table that shows every URL added